GP: Refactor coordinates to float
See: D4030

Differential Revision: https://developer.blender.org/D4036

+++ source/blender/editors/gpencil/gpencil_select.c
@@ -905,8 +905,8 @@ static bool gp_stroke_do_circle_sel(
 			if (((!ELEM(V2D_IS_CLIPPED, x0, y0)) && BLI_rcti_isect_pt(rect, x0, y0)) ||
 			    ((!ELEM(V2D_IS_CLIPPED, x1, y1)) && BLI_rcti_isect_pt(rect, x1, y1)))
 			{
-				int mval[2]  = {mx, my};
-				int mvalo[2] = {mx, my}; /* dummy - this isn't used... */
+				float mval[2]  = {(float)mx, (float)my};
+				float mvalo[2] = {(float)mx, (float)my}; /* dummy - this isn't used... */
 
 				/* check if point segment of stroke had anything to do with
 				 * eraser region  (either within stroke painted, or on its lines)
